feat: add FISH widget
This widget coordinates a fluidic system andimage acquisition.
The fluidic experiment parameters are defined in a separate file
loaded by the widget.
Users can define an ROI for imaging by defining corners of a rectangle
with the Stage Positions box, and can choose to extend the number of
z positions with the z-stack box.