Add fluidics program loading, editing and running
We can now load and visualize the fluidics program in the widget.
Usrs can modify fluidic steps, and select the steps to be run during
the FISH experiment, or independently from image acquisition.
/!\ the `fluidics_control.py` module has a tricky and breaking change:
it now expects rounds to be passed in human convention, with first
round starting at 1, not 0. This is to avoid errors when pausing,
resuming or manually modifying the ongoing experiments.