Add DPC groupBox and preprocessing
We added a separate groupBox to acquire or not DPC images, and to use
them or not for autofocus.
If a camera has too many hot pixels or too much noise, the DPC final
images are not usable. We added the possibility to correct for hot
pixels with a dark image or stack, and to denoise raw DPC images with
gaussian or median filter.
The user can acquire a single frame of DPC (the 4 alternated
illuminations + the 2 resulting images) and an "MDA" stack of DPC images
at several x/y and z positions, also preprocessed for hot pixels and
deboised if wanted.